Use new error link styles in error summary
As part of the notification banner work, we introduced a new govuk-link-style-error mixin to automatically apply error colours to links (including visited and hover styles).

We should apply these styles to the error summary component so things are consistent.
The error summary doesn't currently have visited/hover styles for links, which isn't great, so this change adds those styles.