Add aria-labelledby to alert banners
NOTE: this PR does not update the macro documentation - this will be done separately in https://github.com/alphagov/govuk-frontend/pull/2001

We have decided to keep the aria-labelledby attribute on alert banners too, to make things consistent with the error summary. We were seeing some odd behaviour when testing with NVDA and Voiceover, but this change did not seem to have a negative impact and it also means that when we next revisit the component we're at least working from a place where everything is consistent to start with.